requirements.txt
asgiref==3.6.0
#backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
idna==3.4
Pillow==9.4.0
psycopg2==2.9.5
python-dotenv==0.21.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
